requirements.txt
aiohttp==3.11.14
Flask==3.1.0
google-cloud-aiplatform==1.86.0
google-genai==1.8.0
jsonify==0.5
kafka-python==2.1.3
langchain-core==0.3.49
langchain-google-vertexai==2.0.18
langchain-openai==0.3.11
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
Markdown==3.7
pydantic==2.11.0
pydantic_core==2.33.0
python-dotenv==1.1.0
openai==1.69.0